tailcfg: refactor/implement wire structs for TKA
